requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
apispec==6.6.1
argcomplete==3.3.0
asgiref==3.8.1
attrs==23.2.0
awscli==1.32.19
Babel==2.14.0
beautifulsoup4==4.12.3
blinker==1.8.1
botocore==1.34.19
cachelib==0.9.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
coverage==7.4.0
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==42.0.4
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docutils==0.16
email_validator==2.1.1
entsoe-py==0.6.3
fastapi==0.109.2
flake8==7.0.0
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.2.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.6.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-re2==1.1.20240501
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
gunicorn==22.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.7
imagesize==1.4.1
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
limits==3.11.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.2.0
multidict==6.0.5
multimethod==1.10
mypy-extensions==1.0.0
numpy==1.26.3
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
ordered-set==4.1.0
packaging==24.0
pandas==2.2.2
pandera==0.18.0
pathspec==0.12.1
pendulum==3.0.0
pillow==10.2.0
pluggy==1.5.0
prettytable==3.9.0
prison==0.2.1
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
pyarrow==15.0.0
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pyecharts==2.0.4
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pytest==7.4.4
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.31.1
requests>=2.32.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.10.0
setproctitle==1.3.3
setuptools==66.1.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.30
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
-e git+https://github.com/tur-ium/electricity_lca@86ec103c51659647b3287bd958b40c510a50179d#egg=src
starlette==0.36.3
streamlit==1.31.0
streamlit-echarts==0.4.0
streamlit-vega-lite==0.1.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
typeguard==4.1.5
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.1
uvicorn==0.27.1
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.18.1
